requirements.txt
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
gtts
